DiscreteDomain.size() is now a long integer.  Also simplified treatment of Grids, constraining them to be two-dimensional.  (The Grid2 class is there because of weirdness with the NetBeans Mercurial plugin.  This will be renamed Grid in the next commit.)

+++ core/src/main/java/uk/ac/rdg/resc/edal/coverage/DiscreteCoverage.java
@@ -88,8 +88,7 @@ public interface DiscreteCoverage<P, DO, R> extends Coverage<P, R>
      * Gets the number of distinct values in this coverage.  (Equivalent to
      * {@code getDomain().size()}.)
      * @return the number of distinct values in this coverage.
-     * @todo Do we need this?
      */
-    public int size();
+    public long size();
 
 }
